requirements.txt
pure-python-adb
aiofiles
pyyaml
